backend/requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
openai
python-dotenv
requests
beautifulsoup4
pydub
pocketsphinx